backend/requirements.txt
fastapi
uvicorn
pydantic
python-multipart
google-auth
google-auth-httplib2
google-auth-oauthlib
google-api-python-client
pdfplumber
pandas
python-dotenv
gradio
openai
